+++ spark/src/main/spark-3.x/org/apache/spark/sql/comet/shims/ShimStreamSourceAwareSparkPlan.scala
